avatar processing with jimp

+++ controllers/users/updateAvatar.js
@@ -1,5 +1,6 @@
 const path = require("path");
 const fs = require("fs/promises");
+const Jimp = require("jimp");
 const { User } = require("../../models/user");
 
 const avatarsDir = path.join(__dirname, "../../", "public", "avatars");
@@ -7,20 +8,31 @@ const avatarsDir = path.join(__dirname, "../../", "public", "avatars");
 const updateAvatar = async (req, res) => {
   const { path: tmpUpload, originalname } = req.file;
   const { _id } = req.user;
-  const avatarOriginalname = `${_id}_${originalname}`;
+  const avatarSmallOriginalname = `${_id}_small_${originalname}`;
   try {
     // додаємо id в назву, що файли з однаковою назвою не затирались
-    const resultUpload = path.join(avatarsDir, avatarOriginalname);
-    await fs.rename(tmpUpload, resultUpload);
+    const resultUploadSmall = path.join(avatarsDir, avatarSmallOriginalname);
 
-    const avatarURL = path.join("public", "avatars", avatarOriginalname);
+    await Jimp.read(tmpUpload).then((img) => {
+      return img.resize(250, 250).write(resultUploadSmall);
+    });
+
+    const avatarURL = path.join("public", "avatars", avatarSmallOriginalname);
+
     await User.findByIdAndUpdate(_id, { avatarURL });
 
     res.json({ avatarURL });
   } catch (error) {
-    await fs.unlink(tmpUpload);
+    if (await checkFileExists(tmpUpload)) await fs.unlink(tmpUpload);
     throw error;
   }
 };
+
+function checkFileExists(file) {
+  return fs
+    .access(file, fs.constants.F_OK)
+    .then(() => true)
+    .catch(() => false);
+}
 
 module.exports = updateAvatar;
